split out tests, keep only appender tests
