+++ duckdb_test.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"math"
 	"math/big"
-	"math/rand"
 	"reflect"
 	"testing"
 	"time"
@@ -686,448 +685,6 @@ func TestInterval(t *testing.T) {
 	})
 }
 
-func TestVariousLengthsVARCHAR(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value VARCHAR)`)
-	require.NoError(t, err)
-
-	const maxLength = 1024
-	values := []string{}
-	for i := 0; i <= maxLength; i++ {
-		values = append(values, randString(i))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value string
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, maxLength)
-		assert.Equal(t, values[id], value)
-		assert.Equal(t, len(values[id]), id)
-	}
-}
-
-const (
-	nRandomTestRows = 10000
-)
-
-func TestRandomizedUTINYINT(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value UTINYINT)`)
-	require.NoError(t, err)
-
-	values := []uint8{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, uint8(randInt(0, 255)))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value uint8
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedTINYINT(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value TINYINT)`)
-	require.NoError(t, err)
-
-	values := []int8{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, int8(randInt(-128, 127)))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value int8
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedUSMALLINT(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value USMALLINT)`)
-	require.NoError(t, err)
-
-	values := []uint16{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, uint16(randInt(0, 65535)))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value uint16
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedSMALLINT(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value SMALLINT)`)
-	require.NoError(t, err)
-
-	values := []int16{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, int16(randInt(-32768, 32767)))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value int16
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedUINTEGER(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value UINTEGER)`)
-	require.NoError(t, err)
-
-	values := []uint32{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, uint32(randInt(0, 4294967295)))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value uint32
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedINTEGER(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value INTEGER)`)
-	require.NoError(t, err)
-
-	values := []int32{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, int32(randInt(-2147483648, 2147483647)))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value int32
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedUBIGINT(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value UBIGINT)`)
-	require.NoError(t, err)
-
-	values := []uint64{}
-	for i := 0; i < nRandomTestRows; i++ {
-		v := rand.Uint64()
-		// go sql doesn't support uint64 values with high bit set (see for example https://github.com/lib/pq/issues/72)
-		if v > 9223372036854775807 {
-			v = 9223372036854775807
-		}
-		values = append(values, v)
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value uint64
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedBIGINT(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id BIGINT, value BIGINT)`)
-	require.NoError(t, err)
-
-	values := []int64{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, rand.Int63())
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value int64
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedREAL(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value REAL)`)
-	require.NoError(t, err)
-
-	values := []float32{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, randFloat32(-3.4e+37, 3.4e+37))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value float32
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.True(t, float32similar(values[id], value))
-	}
-}
-
-func TestRandomizedDOUBLE(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value DOUBLE)`)
-	require.NoError(t, err)
-
-	values := []float64{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, randFloat64(-1.6e+307, 1.6e+307))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value float64
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.True(t, float64similar(values[id], value))
-	}
-}
-
-func TestRandomizedVARCHAR(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value VARCHAR)`)
-	require.NoError(t, err)
-
-	values := []string{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, randString(int(randInt(0, 256))))
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value string
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.Equal(t, values[id], value)
-	}
-}
-
-func TestRandomizedTIMESTAMP(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE values(id INTEGER, value TIMESTAMP)`)
-	require.NoError(t, err)
-
-	values := []time.Time{}
-	for i := 0; i < nRandomTestRows; i++ {
-		values = append(values, randTime())
-	}
-
-	for id, value := range values {
-		_, err := db.Exec(`INSERT INTO values VALUES (?, ?)`, id, value)
-		require.NoError(t, err)
-	}
-
-	rows, err := db.Query("SELECT id, value FROM values ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	for rows.Next() {
-		var id int
-		var value time.Time
-		require.NoError(t, rows.Scan(&id, &value))
-		assert.LessOrEqual(t, id, nRandomTestRows)
-		assert.True(t, float32similar(float32(values[id].Unix()), float32(value.Unix())))
-	}
-}
-
-func TestColumnTypeNames(t *testing.T) {
-	t.Parallel()
-	db := openDB(t)
-	defer db.Close()
-
-	_, err := db.Exec(`CREATE TABLE misc(id INTEGER, u8 UTINYINT, i8 TINYINT, u16 USMALLINT, i16 SMALLINT,
-		u32 UINTEGER, i32 INTEGER, u64 UBIGINT, i64 BIGINT, hi HUGEINT,
-		f REAL, d DOUBLE, s VARCHAR, b BOOLEAN)`)
-	require.NoError(t, err)
-
-	_, err = db.Exec(`INSERT INTO misc VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		1, 1, -1, 2, -2, 3, -3, 4, -4, 1234, 5.1, 6.2, "hello", true)
-	require.NoError(t, err)
-
-	rows, err := db.Query("SELECT * FROM misc ORDER BY id ASC")
-	require.NoError(t, err)
-	defer rows.Close()
-
-	expectedDbTypeNames := []string{"INTEGER", "UTINYINT", "TINYINT", "USMALLINT", "SMALLINT",
-		"UINTEGER", "INTEGER", "UBIGINT", "BIGINT", "HUGEINT",
-		"FLOAT", "DOUBLE", "VARCHAR", "BOOLEAN", "TIMESTAMP"}
-	colTypes, err := rows.ColumnTypes()
-	for i, s := range colTypes {
-		assert.NotNil(t, s)
-		assert.Equal(t, expectedDbTypeNames[i], s.DatabaseTypeName())
-	}
-}
-
 func TestEmpty(t *testing.T) {
 	t.Parallel()
 	db := openDB(t)
@@ -1361,51 +918,4 @@ func createTable(db *sql.DB, t *testing.T) *sql.Result {
 	res, err := db.Exec("CREATE TABLE foo(bar VARCHAR, baz INTEGER)")
 	require.NoError(t, err)
 	return &res
-}
-
-const (
-	epsilon32 = 0.0000002
-	epsilon64 = 0.0000000000002
-)
-
-func randInt(lo int64, hi int64) int64 {
-	return rand.Int63n(hi-lo+1) + lo
-}
-
-func randBool() bool {
-	return (rand.Int()%2 == 0)
-}
-
-func randFloat32(lo float32, hi float32) float32 {
-	return (rand.Float32() * (hi - lo)) + lo
-}
-
-func randFloat64(lo float64, hi float64) float64 {
-	return (rand.Float64() * (hi - lo)) + lo
-}
-
-func float32similar(v1 float32, v2 float32) bool {
-	return math.Abs(float64(v2-v1)) < epsilon32
-}
-
-func float64similar(v1 float64, v2 float64) bool {
-	return math.Abs(float64(v2-v1)) < epsilon64
-}
-
-var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
-
-func randString(n int) string {
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = letters[rand.Intn(len(letters))]
-	}
-	return string(b)
-}
-
-func randTime() time.Time {
-	lo := int64(788918400)  // 1995
-	hi := int64(1735689600) // 2025
-	randomEpoch := randInt(lo, hi)
-
-	return time.Unix(randomEpoch, 0)
 }
